Rename persistent dashboard metrics hook
