+++ src/hooks/usePersistentDashboardMetrics.ts
